Mmap SSTableReader and BlockReader
This improves performance quite a bit:

db/fillseq              time:   [9.7127 µs 15.202 µs 22.856 µs]
                        change: [-81.652% -65.104% -23.297%] (p = 0.02 < 0.05)
                        Performance has improved.

db/fillrandom           time:   [9.3320 µs 13.290 µs 18.245 µs]
                        change: [-82.959% -66.628% -27.728%] (p = 0.02 < 0.05)
                        Performance has improved.

db/readseq              time:   [193.32 ns 193.50 ns 193.69 ns]
                        change: [-99.479% -99.305% -99.076%] (p = 0.00 < 0.05)
                        Performance has improved.

db/readrandom           time:   [29.832 µs 29.923 µs 30.014 µs]
                        change: [-96.933% -96.916% -96.898%] (p = 0.00 < 0.05)
                        Performance has improved.

+++ src/db_config.rs
@@ -29,11 +29,11 @@ impl Default for DBConfig {
 }
 
 impl DBConfig {
-    pub fn level_max_size(&self, level: u32) -> u64 {
-        self.level_base_max_size * 10u64.pow(level)
+    pub fn level_max_size(&self, level: u32) -> usize {
+        (self.level_base_max_size * 10u64.pow(level)) as usize
     }
 
-    pub fn level_max_sstables(&self, level: u32) -> u64 {
-        self.level_base_max_sstables * 2_u64.pow(level)
+    pub fn level_max_sstables(&self, level: u32) -> usize {
+        (self.level_base_max_sstables * 2_u64.pow(level)) as usize
     }
 }
